+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 2019.08.22
   - 手动实现XXX：
     - 返回某元素第n层父元素结点
+    - 实现children方法：childrenFn
 - 2019.08.21
   - js知识：
     - Generator